requirements.txt
astroid==3.0.2
cffi==1.16.0
dill==0.3.7
isort==5.13.2
mccabe==0.7.0
platformdirs==4.1.0
pycparser==2.21
pylint==3.0.3
PyNaCl==1.5.0
tomlkit==0.12.3
